fix: change avgScore to float and adjust timeouts
- Change avgScore type from int to float to support decimal averages
- Reduce future.get timeout from 10s to 500ms for faster failure detection
- Increase HTTP timeout from 30s to 60s for image generation
- Improve error logging in map preview generation

+++ plugin/src/main/java/plugin/service/MapRating.java
@@ -46,7 +46,7 @@ public class MapRating {
     @RequiredArgsConstructor
     public static class MapRatingStat {
         public final int totalVotes;
-        public final int avgScore;
+        public final float avgScore;
     }
 
     private static MapRatingData load() {
